requirements.txt
torch
torchvision
flask
flask-cors
pillow
